backend/requirements.txt
# Flask & Server
flask==3.0.0
flask-cors==4.0.0

# Database (use version with wheel for your Python if 4.6 fails)
pymongo>=4.6,<5

# Auth
PyJWT==2.8.0
bcrypt==4.1.2

# File parsing
PyPDF2==3.0.1
python-docx==1.1.0

# ML - Core dependencies
nltk==3.8.1
scikit-learn>=1.2.0

# ML - Advanced models
xgboost>=1.7.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# ML - Data handling (for training)
imbalanced-learn>=0.14.0

# Visualization (optional, for analysis)
matplotlib>=3.7.0
seaborn>=0.12.0

# Utils
python-dotenv==1.0.0
werkzeug==3.0.1
google-generativeai>=0.8.0
